Revert "refactor: split example"
This reverts commit 922e84612b6d955759354cc658ac101205bfda88.
