+++ apps/dashboard/src/app/example/components/example-item/example-item.component.html
