+++ apps/dashboard/src/app/example/components/example-list/example-list.component.html
