+++ apps/dashboard/src/app/example/components/example-list/example-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { ExampleItem } from '../../models/example.item.model';
+import { ExampleItem } from '@app/example/models/example.item.model';
 import { ExampleItemComponent } from '../example-item/example-item.component';
 
 @Component({
