+++ apps/dashboard/src/app/example/example.component.html
