+++ apps/dashboard/src/app/example/example.component.ts
@@ -1,8 +1,8 @@
 import { AsyncPipe } from '@angular/common';
 import { Component, inject, OnInit } from '@angular/core';
 import { ButtonComponent } from '@chirimen-lite-dashboard/ui';
-import { DialogService } from '@shared/service/dialog/dialog.service';
 import { BehaviorSubject, forkJoin } from 'rxjs';
+import { DialogService } from '../shared/service/dialog/dialog.service';
 import { ExampleListComponent } from './components/example-list/example-list.component';
 import { ExampleItem } from './models/example.item.model';
 import { ExampleDialogService } from './services/example.dialog.service';
