+++ apps/dashboard/src/app/example/index.ts
