+++ apps/dashboard/src/app/example/models/example.item.model.ts
