+++ apps/dashboard/src/app/example/models/example.model.ts
