+++ apps/dashboard/src/app/example/services/example.service.ts
