--- libs/example/README.md
@@ -1,7 +1,0 @@
-# lib-example
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test lib-example` to execute the unit tests.
