--- libs/example/src/index.ts
@@ -1,15 +1,0 @@
-// Components
-export * from './lib/components/example-item/example-item.component';
-export * from './lib/components/example-list/example-list.component';
-export * from './lib/example.component';
-
-// Services
-export * from './lib/services/example.dialog.service';
-export * from './lib/services/example.service';
-
-// Models
-export * from './lib/models/example.item.model';
-export * from './lib/models/example.model';
-
-// Functions
-export * from './lib/functions/convert';
