+++ tsconfig.base.json
@@ -14,7 +14,6 @@
     "baseUrl": ".",
     "paths": {
       "@app/*": ["apps/dashboard/src/app/*"],
-      "@chirimen-lite-dashboard/lib-example": ["libs/example/src/index.ts"],
       "@chirimen-lite-dashboard/ui": ["libs/ui/src/index.ts"],
       "@shared/*": ["apps/dashboard/src/app/shared/*"]
     }
